requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0
python-dotenv==1.0.0
flask==3.0.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.3
Pillow==10.0.0
streamlit==1.29.0
chromadb==0.4.22
torch==2.1.0
transformers==4.36.0 